refactor: move BootsTable to /src/components
Move BootsTable to /src/components since we will reuse it on other
pages than TreeDetails
